refactor: move common code to separate functions feat: add ability to filter duplicate submissions
