+++ jba/tests/test_logs/test_tests_analysis.py
@@ -6,7 +6,7 @@ from pandas._testing import assert_frame_equal
 
 from jba.src.models.edu_columns import EduColumnName
 from jba.src.models.edu_logs import TestResult, TestData, TestDataField
-from jba.src.visualization.common import (
+from jba.src.test_logs.analysis import (
     _convert_to_timeline,
     TimelineItem,
     _find_test_result,
